# pkg/goSignals/server/routers.go
/* Package server implements the Stream Management API for OpenID Shared Security Events
 *
 * [OpenID Spec](https://openid.net/specs/openid-sse-framework-1_0.html#management)  HTTP API to be implemented by Event Transmitters. This API can be used by Event Receivers to query and update the Event Stream configuration and status, to add and remove subjects, and to trigger verification.
 *
 * API version: 1.0.0
 * Generated by: Swagger Codegen (https://github.com/swagger-api/swagger-codegen.git)
 */
package server

import (
	"fmt"
	"io/fs"
	"log"
	"net/http"

	"github.com/gorilla/mux"
	"github.com/i2-open/i2goSignals/pkg/goSignals"
	"github.com/prometheus/client_golang/prometheus/promhttp"

	"github.com/i2-open/i2goSignals/pkg/goSSEF"
)

type Route struct {
	Name        string
	Method      string
	Pattern     string
	HandlerFunc http.HandlerFunc
	IsIdQuery   bool
}

type HttpRouter struct {
	router *mux.Router
	sa     *SignalsApplication
}

type Routes []Route

func NewRouter(application *SignalsApplication) *HttpRouter {
	httpRouter := HttpRouter{
		router: mux.NewRouter().StrictSlash(true),
		sa:     application,
	}

	// Add the Prometheus middleware first so logging happens inside
	httpRouter.router.Use(PrometheusHttpMiddleware)

	// Host swagger-ui
	dist, err := fs.Sub(goSignals.SwaggerUI, "swagger-ui")
	if err != nil {
		log.Fatal(err)
	}
	httpRouter.router.PathPrefix("/swagger/").Handler(http.StripPrefix("/swagger/", http.FileServer(http.FS(dist))))

	routes := httpRouter.getRoutes()
	for _, route := range routes {
		var handler http.Handler
		handler = route.HandlerFunc
		handler = application.Logger(handler, route.Name)
		if route.IsIdQuery {
			httpRouter.router.
				Methods(route.Method).
				Path(route.Pattern).
				Name(route.Name).
				Handler(handler).
				Queries("stream_id", "{stream_id:[a-fA-F0-9]+}")
		} else {
			httpRouter.router.
				Methods(route.Method).
				Path(route.Pattern).
				Name(route.Name).
				Handler(handler)
		}

	}

	// Add prometheus handler at metrics endpoint
	httpRouter.router.Path("/metrics").Handler(promhttp.Handler())

	return &httpRouter
}

func (sa *SignalsApplication) Index(w http.ResponseWriter, r *http.Request) {
	test := r.UserAgent()
	_, _ = fmt.Fprintf(w, "Hello %s", test)
}

func (h *HttpRouter) getRoutes() Routes {
	routes := Routes{
		Route{
			"Index",
			"GET",
			"/",
			h.sa.Index,
			false,
		},

		Route{
			"GenerateIat",
			http.MethodGet,
			"/iat",
			h.sa.IssuerProjectIat,
			false,
		},

		Route{
			"RegisterClient",
			http.MethodPost,
			"/register",
			h.sa.RegisterClient,
			false,
		},

		Route{
			"TriggerEvent",
			http.MethodPost,
			"/trigger-event",
			h.sa.TriggerEvent,
			false,
		},

		Route{
			"ReceivePushEvent",
			http.MethodPost,
			"/events/{id}",
			h.sa.ReceivePushEvent,
			false,
		},

		Route{
			"AddSubject",
			http.MethodPost,
			"/add-subject",
			h.sa.AddSubject,
			false,
		},

		Route{
			"GetStatus",
			http.MethodGet,
			"/status",
			h.sa.GetStatus,
			true,
		},

		Route{
			"RemoveSubject",
			http.MethodPost,
			"/remove-subject",
			h.sa.RemoveSubject,
			false,
		},

		Route{
			"StreamDelete",
			http.MethodDelete,
			"/stream",
			h.sa.StreamDelete,
			true,
		},

		Route{"ListStreamStates",
			http.MethodGet,
			"/streams",
			h.sa.ListStreamStates,
			false,
		},

		Route{
			"StreamGet",
			http.MethodGet,
			"/stream",
			h.sa.StreamGet,
			true,
		},

		Route{
			"StreamCreate",
			http.MethodPost,
			"/stream",
			h.sa.StreamCreate,
			false,
		},

		Route{
			"StreamReplace",
			http.MethodPut,
			"/stream",
			h.sa.StreamUpdate,
			false,
		},

		Route{
			"StreamPatch",
			http.MethodPatch,
			"/stream",
			h.sa.StreamUpdate,
			false,
		},

		Route{
			"UpdateStatus",
			http.MethodPost,
			"/status",
			h.sa.UpdateStatus,
			true,
		},

		Route{
			"VerificationRequest",
			http.MethodPost,
			"/verification",
			h.sa.VerificationRequest,
			false,
		},

		Route{
			"WellKnownSsfConfigurationGet",
			http.MethodGet,
			"/.well-known/ssf-configuration",
			h.sa.WellKnownSsfConfigurationGet,
			false,
		},

		Route{
			"WellKnownSsfConfigurationIssuerGet",
			http.MethodGet,
			"/.well-known/ssf-configuration/{issuer}",
			h.sa.WellKnownSsfConfigurationIssuerGet,
			false,
		},
		Route{
			"JwksJsonCreate",
			http.MethodPost,
			"/jwks/{issuer}",
			h.sa.CreateJwksIssuer,
			false,
		},
		Route{
			"JwksJson",
			http.MethodGet,
			"/jwks.json",
			h.sa.JwksJson,
			false,
		},

		Route{
			"JwksJsonTenant",
			http.MethodGet,
			"/jwks/{issuer}",
			h.sa.JwksJsonIssuer,
			false,
		},

		Route{
			"PollEvents",
			http.MethodPost,
			"/poll/{id}",
			h.sa.PollEvents,
			false,
		},

		Route{
			"ProtectedResourceMetadata",
			http.MethodGet,
			"/.well-known/oauth-provisioned-resource",
			h.sa.ProtectedResourceMetadata,
			false,
		},
	}
	return routes
}