# pkg/goSignals/swagger-ui/api/swagger.yaml
openapi: 3.0.3
info:
  title: Stream Management API for OpenID Shared Security Events
  description: |
    [OpenID Spec](https://openid.net/specs/openid-sse-framework-1_0.html#management)

    HTTP API to be implemented by Event Transmitters. This API can be used by
    Event Receivers to query and update the Event Stream configuration and
    status, to add and remove subjects, and to trigger verification.
  license:
    name: MIT
  version: 1.0.0
servers:
- url: /
...
paths:
...
  /stream:
    get:
      tags:
      - StreamManagement
      summary: Request to retrieve the configuration of an event stream
      description: |-
        [Spec](https://openid.net/specs/openid-sse-framework-1_0.html#reading-a-streams-configuration)

        An Event Receiver gets the current configuration of a stream by making an HTTP GET request to the Configuration
        Endpoint. On receiving a valid request the Event Transmitter responds with a 200 OK response containing a JSON
        representation of the stream’s configuration in the body.
      responses:
        "200":
          description: The Event Stream's configuration
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/StreamConfiguration'
        "401":
          description: Request body cannot be parsed or the request is otherwise invalid
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
        "403":
          description: The Event Receiver is not allowed to read the stream configuration.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
        "404":
          description: There is no Event Stream configured for this Event Receiver.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
      security:
      - BearerAuth: []
    post:
      tags:
      - StreamManagement
      summary: Request to update the configuration of an event stream
      description: |-
        [Spec](https://openid.net/specs/openid-sse-framework-1_0.html#updating-a-streams-configuration)

        An Event Receiver updates the current configuration of a stream by making an HTTP POST request to the Configuration
        Endpoint. The POST body contains a JSON representation of the updated configuration. On receiving a valid request
        the Event Transmitter responds with a 200 OK response containing a JSON representation of the updated stream
        configuration in the body.

        The full set of editable properties must be present in the POST body, not only the ones that are specifically
        intended to be changed. Missing properties SHOULD be interpreted as requested to be deleted. Event Receivers should
        read the configuration first, modify the JSON representation, then make an update request.

        Properties that cannot be updated MAY be present, but they MUST match the expected value.
      requestBody:
        description: Request parameters
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/StreamConfiguration'
        required: true
      responses:
        "200":
          description: The Event Stream's updated configuration
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/StreamConfiguration'
        "202":
          description: |-
            The update request has been accepted, but not processed.
            Receiver MAY try the same request later in order to get processing result.

            If a Receiver makes a request to update the stream to enable it for a specific Subject, and the Transmitter is
            unable to decide whether or not to complete the request, then the Transmitter MUST respond with a 202 status
            code.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
        "400":
          description: Request body cannot be parsed or the request is otherwise invalid
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
        "401":
          description: Authorization failed or is missing
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
        "403":
          description: The Event Receiver is not allowed to update the stream configuration.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
        "404":
          description: There is no Event Stream configured for this Event Receiver.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
      security:
      - BearerAuth: []
    delete:
      tags:
      - StreamManagement
      summary: Request to remove the configuration of an event stream
      description: |-
        An Event Receiver removes the configuration of a stream by making an HTTP DELETE request to the Configuration
        Endpoint. On receiving a request the Event Transmitter responds with a 200 OK response if the configuration was
        successfully removed.
      responses:
        "200":
          description: |-
            On receiving a request the Event Transmitter responds with a 200 OK response if the configuration was
            successfully removed.
        "401":
          description: Authorization failed or is missing
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
        "403":
          description: The Event Receiver is not allowed to update the stream configuration.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
      security:
      - BearerAuth: []
  /status:
    get:
      tags:
      - StreamManagement
      summary: Request to get the status of an Event Stream
      description: |-
        [Spec](https://openid.net/specs/openid-sse-framework-1_0.html#reading-a-streams-status)

        An Event Receiver checks the current status of an event stream by making an
        HTTP GET request to the stream’s Status Endpoint.
      operationId: get_status
      parameters:
      - name: subject
        in: query
        description: |-
          OPTIONAL. The subject for which the stream status is requested.
          Content is a string-encoded json representation of a Subject
          (see Subject schema)
        required: false
        style: form
        explode: true
        schema:
          type: string
          example: "{\"user\":{\"format\":\"email\",\"email\":\"reginold@popular-app.com\"\
            }}"
      responses:
        "200":
          description: |-
            On receiving a valid request the Event Transmitter responds with a 200
            OK response containing a JSON object with an attribute status
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/StreamStatus'
        "401":
          description: Authorization failed or is missing
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
        "403":
          description: The Event Receiver is not allowed to read this stream's status.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
        "404":
          description: |-
            There is no Event Stream configured for this Event Receiver,
            or the Subject specified is invalid or
            the Receiver is not authorized to get status for the specified Subject.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
      security:
      - BearerAuth: []
    post:
      tags:
      - StreamManagement
      summary: Request to update an Event Stream's status
      description: |-
        [Spec](https://openid.net/specs/openid-sse-framework-1_0-ID1.html#updating-a-streams-status)

        An Event Receiver updates the current status of a stream by making an HTTP
        POST request to the Status Endpoint.
      operationId: update_status
      requestBody:
        description: Request parameters
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/UpdateStreamStatus'
        required: true
      responses:
        "200":
          description: Event Stream status successfully updated
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/UpdateStreamStatus'
        "202":
          description: |-
            The update request has been accepted, but not processed.
            Receiver MAY try the same request later in order to get processing result.
        "400":
          description: Request body cannot be parsed or the request is otherwise invalid
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
        "401":
          description: Authorization failed or is missing
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
        "403":
          description: The Event Receiver is not allowed to update the stream status.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
        "404":
          description: |-
            There is no Event Stream configured for this Event Receiver,
            or an invalid Subject is specified.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
      security:
      - BearerAuth: []
  /add-subject:
    post:
      tags:
      - StreamManagement
      summary: Request to add a subject to an Event Stream
      description: |-
        [Spec](https://openid.net/specs/openid-sse-framework-1_0.html#adding-a-subject-to-a-stream)

        Event Receivers can send requests to an Event Transmitter's Add Subject endpoint to add a subject to an Event Stream.
      operationId: add_subject
      requestBody:
        description: Request parameters
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/AddSubjectParameters'
        required: true
      responses:
        "200":
          description: |-
            On a successful response, the Event Transmitter responds with an empty 200 OK response.
            The Event Transmitter MAY choose to silently ignore the request, for example if the subject has previously
            indicated to the transmitter that they do not want events to be transmitted to the Event Receiver.
            In this case, the transmitter MAY return an empty 200 OK response or an appropriate error code.
            See [Security Considerations](https://openid.net/specs/openid-sse-framework-1_0.html#management-sec).
        "400":
          description: Request body cannot be parsed or the request is otherwise invalid
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
        "401":
          description: Authorization failed or is missing
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
        "403":
          description: "The Event Receiver is not allowed to add this particular subject,\
            \ or not allowed to add in general."
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
        "404":
          description: |-
            The subject is not recognized by the Event Transmitter,
            the Event Transmitter may chose to stay silent in this case and respond with 200.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
        "429":
          description: Event Receiver is sending too many requests
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
      security:
      - BearerAuth: []
  /remove-subject:
    post:
      tags:
      - StreamManagement
      summary: Request to add a subject to an Event Stream
      description: |-
        [Spec](https://openid.net/specs/openid-sse-framework-1_0.html#removing-a-subject)

        Event Receivers can send requests to an Event Transmitter's Remove Subject endpoint to remove a subject from an Event Stream.
      operationId: remove_subject
      requestBody:
        description: Request parameters
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/RemoveSubjectParameters'
        required: true
      responses:
        "204":
          description: "On a successful response, the Event Transmitter responds with\
            \ a 204 No Content response."
        "400":
          description: Request body cannot be parsed or the request is otherwise invalid
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
        "401":
          description: Authorization failed or is missing
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
        "403":
          description: "The Event Receiver is not allowed to remove this particular\
            \ subject, or not allowed to remove in general."
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
        "404":
          description: |-
            The subject is not recognized by the Event Transmitter,
            the Event Transmitter may chose to stay silent in this case and respond with 204.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
        "429":
          description: Event Receiver is sending too many requests
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
      security:
      - BearerAuth: []
  /verification:
    post:
      tags:
      - StreamManagement
      summary: Request that a verification event be sent over an Event Stream
      operationId: verification_request
      requestBody:
        description: Optional request parameters
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/VerificationParameters'
        required: false
      responses:
        "204":
          description: |
            Request for verification event was successfully received. A
            successful response from a POST to the Verification Endpoint does
            not indicate that the verification event was transmitted
            successfully, only that the Event Transmitter has transmitted the
            event or will do so at some point in the future.
        "400":
          description: Request body cannot be parsed or the request is otherwise invalid
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
        "401":
          description: Authorization failed or is missing
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
        "429":
          description: Event Receiver is sending too many requests
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
      security:
      - BearerAuth: []
...
  /register:
    post:
      tags:
      - OutOfBand
      summary: Request the transmitter to setup a stream for a given receiver.
      description: |-
        This endpoint is not part of the spec, but rather a quick-and-dirty way to allow a receiver to pass along its
        audience claim in exchange for a bearer token.
      operationId: register
      requestBody:
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/RegisterParameters'
        required: true
      responses:
        "200":
          description: "On successful creation of a stream, returns a bearer token\
            \ that will be used for subsequent requests."
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/RegisterResponse'
  /trigger-event:
    post:
      tags:
      - OutOfBand
      summary: Request the transmitter to create a SSE event of certain type and subject
        and send it to the (streams)receivers that care about the subject.
      description: "This endpoint is not part of the spec, but rather a quick-and-dirty\
        \ way to  test out sample security events other than verification."
      operationId: trigger_event
      requestBody:
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/TriggerEventParameters'
        required: true
      responses:
        "200":
          description: "On successful creation of an event, it will be sent out as\
            \ per SSE spec (i.e. push or poll)."
components:
  schemas:
    StreamStatus:
      required:
      - status
      type: object
      properties:
        status:
          type: string
          description: |-
            REQUIRED. The status of the stream. Values can be one of:

            enabled:
              The Transmitter MUST transmit events over the stream,
              according to the stream’s configured delivery method.

            paused:
              The Transmitter MUST NOT transmit events over the stream.
              The transmitter will hold any events it would have transmitted while paused,
              and SHOULD transmit them when the stream’s status becomes enabled.
              If a Transmitter holds successive events that affect the same Subject Principal,
              then the Transmitter MUST make sure that those events are transmitted in
              the order of time that they were generated OR the Transmitter MUST send
              only the last events that do not require the previous events affecting
              the same Subject Principal to be processed by the Receiver,
              because the previous events are either cancelled by the later events or
              the previous events are outdated.

            disabled:
              The Transmitter MUST NOT transmit events over the stream,
              and will not hold any events for later transmission.
          example: disabled
          enum:
          - enabled
          - paused
          - disabled
        subject:
          description: OPTIONAL. The Subject to which the status applies.
          example:
            format: email
            email: reginold@popular-app.com
          allOf:
          - $ref: '#/components/schemas/Subject'
      example:
        subject:
          format: email
          email: reginold@popular-app.com
        status: disabled
...
    StreamConfiguration:
      title: Event Stream Configuration
      required:
      - delivery
      - events_requested
      type: object
      properties:
        iss:
          type: string
          description: |-
            Read-Only.
            A URL using the https scheme with no query or fragment component that the Transmitter asserts as its Issuer
            Identifier. This MUST be identical to the iss Claim value in Security Event Tokens issued from this Transmitter.
          format: uri
          example: https://most-secure.com
        aud:
          description: |-
            Read-Only.
            A string or an array of strings containing an audience claim as defined in
            [JSON Web Token (JWT)](https://openid.net/specs/openid-sse-framework-1_0.html#RFC7519) that identifies
            the Event Receiver(s) for the Event Stream. This property cannot be updated. If multiple Receivers are specified
            then the Transmitter SHOULD know that these Receivers are the same entity.
          example: https://popular-app.com
          oneOf:
          - type: string
            format: uri
          - type: array
            items:
              type: string
              format: uri
        events_supported:
          type: array
          description: |-
            Read-Only.
            An array of URIs identifying the set of events supported by the Transmitter for this Receiver.
            If omitted, Event Transmitters SHOULD make this set available to the Event Receiver via some other means
            (e.g. publishing it in online documentation).
          example:
          - https://schemas.openid.net/secevent/caep/event-type/session-revoked
          - https://schemas.openid.net/secevent/risc/event-type/credential-compromise
          items:
            type: string
            format: uri
        events_requested:
          type: array
          description: |-
            Read-Write.
            An array of URIs identifying the set of events that the Receiver requested.
            A Receiver SHOULD request only the events that it understands and it can act on.
            This is configurable by the Receiver.
          example:
          - https://schemas.openid.net/secevent/risc/event-type/credential-compromise
          items:
            type: string
            format: uri
        events_delivered:
          type: array
          description: |-
            Read-Only.
            An array of URIs which is the intersection of events_supported and events_requested.
            These events MAY be delivered over the Event Stream.
          example:
          - https://schemas.openid.net/secevent/risc/event-type/credential-compromise
          items:
            type: string
            format: uri
        delivery:
          description: |-
            Read-Write.
            A JSON object containing a set of name/value pairs specifying configuration parameters for the SET delivery
            method. The actual delivery method is identified by the special key method with the value being a URI as defined
            in [Section 11.2.1](https://openid.net/specs/openid-sse-framework-1_0.html#delivery-meta).
          example:
            method: https://schemas.openid.net/secevent/risc/delivery-method/poll
            endpoint_url: null
          discriminator:
            propertyName: method
          oneOf:
          - $ref: '#/components/schemas/PushDeliveryMethod'
          - $ref: '#/components/schemas/PollDeliveryMethod'
        min_verification_interval:
          type: integer
          description: |-
            Read-Only.
            An integer indicating the minimum amount of time in seconds that must pass in between verification requests.
            If an Event Receiver submits verification requests more frequently than this, the Event Transmitter MAY respond
            with a 429 status code. An Event Transmitter SHOULD NOT respond with a 429 status code if an Event Receiver is not
            exceeding this frequency.
        format:
          type: string
          description: |-
            Read-Write.
            The Subject Identifier Format that the Receiver wants for the events.
            If not set then the Transmitter might decide to use a type that discloses more information than necessary.
      description: |
        JSON Object describing and Event Stream's configuration
        [Spec](https://openid.net/specs/openid-sse-framework-1_0.html#stream-config)"
      example:
        aud: https://popular-app.com
        delivery:
          method: https://schemas.openid.net/secevent/risc/delivery-method/poll
          endpoint_url: null
        events_supported:
        - https://schemas.openid.net/secevent/caep/event-type/session-revoked
        - https://schemas.openid.net/secevent/risc/event-type/credential-compromise
        iss: https://most-secure.com
        format: format
        events_delivered:
        - https://schemas.openid.net/secevent/risc/event-type/credential-compromise
        events_requested:
        - https://schemas.openid.net/secevent/risc/event-type/credential-compromise
        min_verification_interval: 0
    UpdateStreamStatus:
      type: object
      properties:
        reason:
          type: string
          description: OPTIONAL. A short text description that explains the reason
            for the change.
          example: Disabled by administrator action.
      example:
        reason: Disabled by administrator action.
      allOf:
      - $ref: '#/components/schemas/StreamStatus'
    PollDeliveryMethod:
      required:
      - method
      type: object
      properties:
        method:
          type: string
          default: https://schemas.openid.net/secevent/risc/delivery-method/poll
          enum:
          - https://schemas.openid.net/secevent/risc/delivery-method/poll
        endpoint_url:
          type: string
          description: The URL where events can be retrieved from. This is specified
            by the Transmitter.
          format: uri
          nullable: true
      example:
        method: https://schemas.openid.net/secevent/risc/delivery-method/poll
        endpoint_url: https://transmitter.most-secure.com/poll
    PushDeliveryMethod:
      required:
      - endpoint_url
      - method
      type: object
      properties:
        method:
          type: string
          default: https://schemas.openid.net/secevent/risc/delivery-method/push
          enum:
          - https://schemas.openid.net/secevent/risc/delivery-method/push
        endpoint_url:
          type: string
          description: The URL where events are pushed through HTTP POST. This is
            set by the Receiver.
          format: uri
        authorization_header:
          type: string
          description: |-
            The HTTP Authorization header that the Transmitter MUST set with each event delivery,
            if the configuration is present. The value is optional and it is set by the Receiver.
      example:
        method: https://schemas.openid.net/secevent/risc/delivery-method/push
        endpoint_url: https://receiver.example.com/push
    RegisterResponse:
      required:
      - token
      type: object
      properties:
        token:
          type: string
          description: The Bearer Token that the receiver will need to use for all
            Stream Management API calls that require authorization.
      example:
        token: 49e5e7785e4e4f688aa49e2585970370
    TriggerEventParameters:
      title: Trigger Event Parameters
      required:
      - event_type
      - subject
      type: object
      properties:
        event_type:
          type: string
          description: "Supports all [RISC](https://openid.net/specs/openid-risc-profile-specification-1_0-01.html)\
            \ and [CAEP](https://openid.net/specs/openid-caep-specification-1_0-ID1.html)\
            \ event types."
          example: credential-compromise
          enum:
          - session-revoked
          - token-claims-change
          - credential-change
          - assurance-level-change
          - device-compliance-change
          - account-purged
          - account-disabled
          - account-enabled
          - identifier-changed
          - identifier-recycled
          - credential-compromise
          - opt-in
          - opt-out-initiated
          - opt-out-cancelled
          - opt-out-effective
          - recovery-activated
          - recovery-information-changed
        subject:
          allOf:
          - $ref: '#/components/schemas/Subject'
          - description: |
              REQUIRED. ONLY EMAIL format supported currently.
              SHOULD be same as subjects receiver cares about.
              A Subject claim identifying the subject of the event to be generated.
      description: "JSON Object describing request to create a security event to test\
        \ SSE receiver/transmitter \n"
      example:
        event_type: session-revoked
        subject:
          format: email
          email: user@example.com
    AddSubjectParameters:
      required:
      - subject
      type: object
      properties:
        subject:
          allOf:
          - $ref: '#/components/schemas/Subject'
          - description: REQUIRED. A Subject claim identifying the subject to be added.
        verified:
          type: boolean
          description: |-
            OPTIONAL. A boolean value; when true, it indicates that the Event Receiver has verified the Subject claim.
            When false, it indicates that the Event Receiver has not verified the Subject claim.
            If omitted, Event Transmitters SHOULD assume that the subject has been verified.
      example:
        subject:
          format: email
          email: reginold@popular-app.com
        verified: true
...
    RegisterParameters:
      type: object
      properties:
        audience:
          type: string
          description: The audience claim to be used for all events on this stream.
          format: uri
          example: https://popular-app.com
    RemoveSubjectParameters:
      required:
      - subject
      type: object
      properties:
        subject:
          allOf:
          - $ref: '#/components/schemas/Subject'
          - description: REQUIRED. A Subject claim identifying the subject to be removed.
      example:
        subject:
          format: email
          email: reginold@popular-app.com
    VerificationParameters:
      type: object
      properties:
        state:
          type: string
          description: |
            OPTIONAL. An arbitrary string that the Event Transmitter
            MUST echo back to the Event Receiver in the verification
            event’s payload. Event Receivers MAY use the value of this
            parameter to correlate a verification event with a
            verification request. If the verification event is
            initiated by the transmitter then this parameter MUST not
            be set.
...
    Error:
      required:
      - code
      - message
      type: object
      properties:
        code:
          type: string
        message:
          type: string
...
  securitySchemes:
    BearerAuth:
      type: http
      scheme: bearer